// src/primer-api/model/paginatedMeta.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * Primer backend API
 * A backend service implementing a pedagogic functional programming language.
 * OpenAPI spec version: 0.7
 */

export interface PaginatedMeta {
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   * @exclusiveMinimum
   */
  firstPage: number;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   * @exclusiveMinimum
   */
  lastPage: number;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   * @exclusiveMinimum
   */
  nextPage?: number;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   * @exclusiveMinimum
   */
  pageSize: number;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   * @exclusiveMinimum
   */
  prevPage?: number;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   * @exclusiveMinimum
   */
  thisPage: number;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   */
  totalItems: number;
}